update booking details

+++ src/bookings/booking.controller.ts
@@ -3,7 +3,9 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
   Param,
+  Patch,
   Post,
   Req,
   Request,
@@ -18,6 +20,8 @@ import { Roles } from 'src/auth/roles.decorator';
 import { RolesGuard } from 'src/auth/roles.guard';
 import { Role } from 'src/auth/roles.enum';
 import { Response } from 'express';
+import {updateBookingDto} from 'src/bookings/dto/UpdateBookingRequest.dto'
+import mongoose from 'mongoose';
 
 @Controller('booking')
 export class BookingController {
@@ -54,11 +58,35 @@ export class BookingController {
       return await this.bookingService.getBookingsForUser(tenantId)
     }
 
+    @Patch(':id')
+    @UseGuards(JwtAuthGuard, RolesGuard)
+    @Roles(Role.Tenant)
+    async updateBookings(@Param('id') bookingId: string, @Body() updateBookingDto: updateBookingDto,  @Req() Request: CustomRequest){
+      const tenantId = Request.user.id
+      if(!tenantId){
+        throw new HttpException('tenant not found', 404)
+      }
+      const isValid = mongoose.Types.ObjectId.isValid(bookingId);
+    if (!bookingId) {
+      throw new HttpException('invalid booking id', 400);
+    }
+
+    const updateBookig = await this.bookingService.updateBookingsForUser(updateBookingDto, bookingId, tenantId);
+    return updateBookig
+    }
+
     @Delete(':id')
     @UseGuards(JwtAuthGuard, RolesGuard)
     @Roles(Role.Tenant)
     async deleteBookingForUser(@Param('id') bookingId: string, @Req() Request: CustomRequest){
       const tenantId = Request.user.id
+      if(!tenantId){
+        throw new HttpException('tenant not found', 404)
+      }
+      if (!bookingId) {
+        throw new HttpException('invalid booking id', 400);
+      }
+  
       return await this.bookingService.deleteBookingForUser(bookingId, tenantId)
     }
 }
